+++ src/DSP/samplers/Sampler.h
@@ -100,16 +100,16 @@ private:
     InputNode input_start_mod;
     OutputNode output;
     
-    float readIndex;
-    float inc;
+    double readIndex;
+    double inc;
     SampleBuffer* sample;
     int channel;
     int sampleIndex;
     
     std::vector<SampleBuffer*> samples;
     
-    float incBase;
-    float direction;
+    double incBase;
+    double direction;
     
     std::atomic<float> positionMeter;
     float positionDivider;
